+++ src/server/scripts/Northrend/Ulduar/Ulduar/boss_xt002.cpp
@@ -700,8 +700,8 @@ class npc_boombot : public CreatureScript
                     _boomed = true; // Prevent recursive calls
 
                     WorldPacket data(SMSG_SPELLINSTAKILLLOG, 8+8+4);
-                    data << uint64(me->GetGUID());
-                    data << uint64(me->GetGUID());
+                    data << me->GetGUID();
+                    data << me->GetGUID();
                     data << uint32(SPELL_BOOM);
                     me->SendMessageToSet(&data, false);
